Small fixes (#32)
* fix docker build warning

prior to this, building the image was causing this warning:

```
 1 warning found (use docker --debug to expand):
 - FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 2)
```

* dockerfiles: cache go mod dependencies in their own layer

this change causes docker to create an additional intermediate layer containing
just the go.mod file and the downloaded dependencies cited therefrom. this
prevents having to re-download the go.mod dependencies on EVERY image rebuild.
more specifically, as long as you don't change the go.mod file, docker will use
the cached layer.

* README: dev dependencies: add gci

it's used in the Makefile but, prior to this, the README didn't instruct you to
install it.